Deleted unnecessary dotfiles folder
Moved all files into root folder of repository and fixed set-from-git to reflect
that. I was able to get set-from-git to work without listing all the dotfiles,
but it didn't work as planned due to swp files, git folder and others; it's just
easier and cleaner to list out the files.